refac general improvements
